Format files with ruff

+++ api/src/core/auth.py
@@ -20,28 +20,31 @@ from service.error import TooManyRequests, ApiError, NotFound, Unauthorized, Bad
 from typing import Optional
 
 
-logger = getLogger('makeradmin')
+logger = getLogger("makeradmin")
 
 
 def generate_token() -> str:
-    return ''.join(secrets.choice(ascii_letters + digits) for _ in range(32))
+    return "".join(secrets.choice(ascii_letters + digits) for _ in range(32))
 
 
 def get_member_by_user_identification(user_identification):
     try:
         if user_identification.isdigit():
             return db_session.query(Member).filter_by(member_number=int(user_identification), deleted_at=None).one()
-            
+
         return db_session.query(Member).filter_by(email=user_identification, deleted_at=None).one()
-        
+
     except NoResultFound:
-        raise NotFound(f"Could not find any user with the name or email '{user_identification}'.",
-                       fields='user_identification', status="not found")
+        raise NotFound(
+            f"Could not find any user with the name or email '{user_identification}'.",
+            fields="user_identification",
+            status="not found",
+        )
 
 
-def create_access_token(ip: str, browser: Optional[str], user_id: int, valid_duration: Optional[timedelta]=None):
+def create_access_token(ip: str, browser: Optional[str], user_id: int, valid_duration: Optional[timedelta] = None):
     assert user_id > 0
-    
+
     access_token = AccessToken(
         user_id=user_id,
         access_token=generate_token(),
@@ -50,26 +53,27 @@ def create_access_token(ip: str, browser: Optional[str], user_id: int, valid_dur
         expires=datetime.utcnow() + (timedelta(minutes=15) if valid_duration is None else valid_duration),
         lifetime=int(timedelta(days=14).total_seconds()),
     )
-    
+
     db_session.add(access_token)
-    
+
     return dict(access_token=access_token.access_token, expires=access_token.expires.isoformat())
 
 
 def login(ip, browser, username, password):
-    
     count = Login.get_failed_login_count(ip)
-    
+
     if count > 10:
-        raise TooManyRequests("Your have reached your maximum number of failed login attempts for the last hour."
-                              " Please try again later.")
+        raise TooManyRequests(
+            "Your have reached your maximum number of failed login attempts for the last hour."
+            " Please try again later."
+        )
 
     try:
         member_id = authenticate(username=username, password=password)
     except ApiError:
         Login.register_login_failed(ip)
         raise
-    
+
     Login.register_login_success(ip, member_id)
 
     return create_access_token(ip, browser, member_id)
@@ -77,14 +81,15 @@ def login(ip, browser, username, password):
 
 def request_password_reset(user_identification):
     member = get_member_by_user_identification(user_identification)
-    
+
     token = generate_token()
-    
+
     db_session.add(PasswordResetToken(member_id=member.member_id, token=token))
     db_session.flush()
-    
+
     send_message(
-        MessageTemplate.PASSWORD_RESET, member,
+        MessageTemplate.PASSWORD_RESET,
+        member,
         url=config.get_admin_url(f"/password-reset?reset_token={quote_plus(token)}"),
     )
 
@@ -92,31 +97,31 @@ def request_password_reset(user_identification):
 def password_reset(reset_token, unhashed_password):
     try:
         password_reset_token = db_session.query(PasswordResetToken).filter_by(token=reset_token).one()
-        
+
     except NoResultFound:
         return dict(error_message="Could not find password reset token, try to request a new reset link.")
-    
+
     except MultipleResultsFound:
         raise InternalServerError(log=f"Multiple tokens {reset_token} found, this is a bug.")
-    
+
     if datetime.utcnow() - password_reset_token.created_at > timedelta(minutes=10):
         return dict(error_message="Reset link expired, try to request a new.")
-    
+
     try:
         hashed_password = check_and_hash_password(unhashed_password)
     except ValueError as e:
         return dict(error_message=str(e))
-    
+
     try:
         member = db_session.query(Member).get(password_reset_token.member_id)
     except NoResultFound:
         raise InternalServerError(log=f"No member with id {password_reset_token.member_id} found, this is a bug.")
-    
+
     member.password = hashed_password
     db_session.add(member)
-    
+
     return {}
-    
+
 
 def force_login(ip: str, browser: str, user_id: int):
     Login.register_login_success(ip, user_id)
@@ -125,29 +130,31 @@ def force_login(ip: str, browser: str, user_id: int):
 
 def remove_token(token, user_id):
     assert user_id > 0
-    
-    count = db_session\
-        .query(AccessToken)\
-        .filter(AccessToken.user_id == user_id, AccessToken.access_token == token)\
-        .delete()
-    
+
+    count = (
+        db_session.query(AccessToken).filter(AccessToken.user_id == user_id, AccessToken.access_token == token).delete()
+    )
+
     if not count:
         raise NotFound("The access_token you specified could not be found in the database.")
-        
+
     return None
 
 
 def list_for_user(user_id):
-    return [dict(
-        access_token=access_token.access_token,
-        browser=access_token.browser,
-        ip=access_token.ip,
-        expires=access_token.expires.isoformat(),
-    ) for access_token in db_session.query(AccessToken).filter(AccessToken.user_id == user_id)]
+    return [
+        dict(
+            access_token=access_token.access_token,
+            browser=access_token.browser,
+            ip=access_token.ip,
+            expires=access_token.expires.isoformat(),
+        )
+        for access_token in db_session.query(AccessToken).filter(AccessToken.user_id == user_id)
+    ]
 
 
 def authenticate_request() -> None:
-    """ Update global object with user_id and user permissions using token from request header. """
+    """Update global object with user_id and user permissions using token from request header."""
 
     # Make sure user_id and permissions is always set.
     g.user_id = None
@@ -158,49 +165,50 @@ def authenticate_request() -> None:
     # logger.info("ARGS " + repr(request.args))
     # logger.info("FORM " + repr(request.form))
     # logger.info("JSON " + repr(request.json))
-    
-    authorization = request.headers.get('Authorization', None)
+
+    authorization = request.headers.get("Authorization", None)
     if authorization is None:
         return
 
-    bearer = 'Bearer '
+    bearer = "Bearer "
     if not authorization.startswith(bearer):
         raise Unauthorized("Unauthorized, can't find credentials.", fields="bearer", what=REQUIRED)
-        
-    token = authorization[len(bearer):].strip()
-    
+
+    token = authorization[len(bearer) :].strip()
+
     access_token = db_session.query(AccessToken).get(token)
     if not access_token:
         raise Unauthorized("Unauthorized, invalid access token.", fields="bearer", what=BAD_VALUE)
-    
+
     now = datetime.utcnow()
     if access_token.expires < now:
         db_session.query(AccessToken).filter(AccessToken.expires < now).delete()
         raise Unauthorized("Unauthorized, expired access token.", fields="bearer", what=EXPIRED)
-    
+
     if access_token.permissions is None:
         if access_token.user_id < 0:
             permissions = SERVICE_PERMISSIONS.get(access_token.user_id, [])
-            
+
         elif access_token.user_id > 0:
             permissions_set = {p for _, p in get_member_permissions(access_token.user_id)}
             permissions_set.add(USER)
             permissions = list(permissions_set)
-            
+
         else:
-            raise BadRequest("Bad token.",
-                             log=f"access_token {access_token.access_token} has user_id 0, this should never happend")
-            
-        access_token.permissions = ','.join(permissions)
-    
+            raise BadRequest(
+                "Bad token.", log=f"access_token {access_token.access_token} has user_id 0, this should never happend"
+            )
+
+        access_token.permissions = ",".join(permissions)
+
     access_token.ip = request.remote_addr
     access_token.browser = request.user_agent.string
     access_token.expires = datetime.utcnow() + timedelta(seconds=access_token.lifetime)
-    
+
     g.user_id = access_token.user_id
     g.session_token = access_token.access_token
-    g.permissions = access_token.permissions.split(',')
-    
+    g.permissions = access_token.permissions.split(",")
+
     # Commit token validation to make it stick even if request fails later.
     db_session.commit()
 
@@ -218,9 +226,12 @@ def roll_service_token(user_id):
 
 
 def list_service_tokens():
-    return [dict(
-        user_id=access_token.user_id,
-        service_name=SERVICE_NAMES.get(access_token.user_id, "unknown service"),
-        access_token=access_token.access_token,
-        permissions=",".join(SERVICE_PERMISSIONS.get(access_token.user_id, [])),
-    ) for access_token in db_session.query(AccessToken).filter(AccessToken.user_id < 0)]
+    return [
+        dict(
+            user_id=access_token.user_id,
+            service_name=SERVICE_NAMES.get(access_token.user_id, "unknown service"),
+            access_token=access_token.access_token,
+            permissions=",".join(SERVICE_PERMISSIONS.get(access_token.user_id, [])),
+        )
+        for access_token in db_session.query(AccessToken).filter(AccessToken.user_id < 0)
+    ]
